Implement discover API, with proposed extensions
The discover API will always return the environment variables needed
to create the credentials required for --initialize.

--connection-url in turn will always return an array of objects each
consisting of a unique identifier, a connection url, a corresponding
internal port for that url and a default flag to indicate if it should
be the default port to bind to when doing a db tunnel.

See aptible/sweetness#689 and aptible/api.aptible.com#441